feat: add style of main page
